Add ECPMasterPattern class derived from abstract Kikuchi MP class
