+++ kikuchipy/signals/ebsd_master_pattern.py
@@ -16,23 +16,25 @@
 # You should have received a copy of the GNU General Public License
 # along with kikuchipy. If not, see <http://www.gnu.org/licenses/>.
 
-import copy
 import gc
 import sys
-from typing import Tuple, Union
+from typing import Union
 
 import dask.array as da
 from dask.diagnostics import ProgressBar
-from hyperspy._lazy_signals import LazySignal2D
 from hyperspy._signals.signal2d import Signal2D
 import numpy as np
-from orix.crystal_map import CrystalMap, Phase, PhaseList
+from orix.crystal_map import CrystalMap, PhaseList
 from orix.quaternion import Rotation
 from skimage.util.dtype import dtype_range
 
+from kikuchipy.signals._kikuchi_master_pattern import (
+    KikuchiMasterPattern,
+    LazyKikuchiMasterPattern,
+)
+
 from kikuchipy.detectors.ebsd_detector import EBSDDetector
 from kikuchipy.signals import LazyEBSD, EBSD
-from kikuchipy.signals._common_image import CommonImage
 from kikuchipy.signals.util._dask import get_chunking
 from kikuchipy.signals.util._master_pattern import (
     _get_direction_cosines_for_single_pc_from_detector,
@@ -40,7 +42,7 @@ from kikuchipy.signals.util._master_pattern import (
 )
 
 
-class EBSDMasterPattern(CommonImage, Signal2D):
+class EBSDMasterPattern(KikuchiMasterPattern, Signal2D):
     """Simulated Electron Backscatter Diffraction (EBSD) master pattern.
 
     This class extends HyperSpy's Signal2D class for EBSD master
@@ -48,40 +50,12 @@ class EBSDMasterPattern(CommonImage, Signal2D):
     HyperSpy user guide. See the docstring of
     :class:`hyperspy.signal.BaseSignal` for a list of additional
     attributes.
-
-    Attributes
-    ----------
-    projection : str
-        Which projection the pattern is in, "stereographic" or
-        "lambert".
-    hemisphere : str
-        Which hemisphere the data contains: "north", "south" or "both".
-    phase : orix.crystal_map.phase_list.Phase
-        Phase describing the crystal structure used in the master
-        pattern simulation.
     """
 
     _signal_type = "EBSDMasterPattern"
     _alias_signal_types = ["ebsd_master_pattern", "master_pattern"]
-    _lazy = False
 
     # ---------------------- Custom properties ----------------------- #
-
-    phase = Phase()
-    projection = None
-    hemisphere = None
-
-    def __init__(self, *args, **kwargs):
-        """Create an :class:`~kikuchipy.signals.EBSDMasterPattern`
-        instance from a :class:`hyperspy.signals.Signal2D` or a
-        :class:`numpy.ndarray`. See the docstring of
-        :class:`hyperspy.signal.BaseSignal` for optional input
-        parameters.
-        """
-        Signal2D.__init__(self, *args, **kwargs)
-        self.phase = kwargs.pop("phase", Phase())
-        self.projection = kwargs.pop("projection", None)
-        self.hemisphere = kwargs.pop("hemisphere", None)
 
     def get_patterns(
         self,
@@ -258,77 +232,8 @@ class EBSDMasterPattern(CommonImage, Signal2D):
 
         return out
 
-    # ------ Methods overwritten from hyperspy.signals.Signal2D ------ #
 
-    def deepcopy(self):
-        new = super().deepcopy()
-        new.phase = self.phase.deepcopy()
-        new.projection = copy.deepcopy(self.projection)
-        new.hemisphere = copy.deepcopy(self.hemisphere)
-        return new
-
-    # ------------------------ Private methods ----------------------- #
-
-    def _get_master_pattern_arrays_from_energy(
-        self, energy: Union[int, float]
-    ) -> Tuple[np.ndarray, np.ndarray]:
-        """Return northern and southern master patterns created with a
-        single, given energy.
-
-        Parameters
-        ----------
-        energy
-            Acceleration voltage in kV. If only a single energy is
-            present in the signal, this will be returned no matter its
-            energy.
-
-        Returns
-        -------
-        master_north, master_south
-            Northern and southern hemispheres of master pattern.
-        """
-        if "energy" in [i.name for i in self.axes_manager.navigation_axes]:
-            master_patterns = self.inav[float(energy)].data
-        else:  # Assume a single energy
-            master_patterns = self.data
-        if self.hemisphere == "both":
-            master_north, master_south = master_patterns
-        else:
-            master_north = master_south = master_patterns
-        return master_north, master_south
-
-    def _is_suitable_for_projection(self, raise_if_not: bool = False):
-        """Check whether the master pattern is suitable for projection
-        onto an EBSD detector and return a bool or raise an error
-        message if desired.
-        """
-        suitable = True
-        error = None
-        if self.projection != "lambert":
-            error = NotImplementedError(
-                "Master pattern must be in the square Lambert projection"
-            )
-            suitable = False
-        pg = self.phase.point_group
-        if pg is None:
-            error = AttributeError(
-                "Master pattern `phase` attribute must have a valid point group"
-            )
-            suitable = False
-        elif self.hemisphere != "both" and not pg.contains_inversion:
-            error = AttributeError(
-                "For point groups without inversion symmetry, like the current "
-                f"{pg.name}, both hemispheres must be present in the master pattern "
-                "signal"
-            )
-            suitable = False
-        if not suitable and raise_if_not:
-            raise error
-        else:
-            return suitable
-
-
-class LazyEBSDMasterPattern(EBSDMasterPattern, LazySignal2D):
+class LazyEBSDMasterPattern(LazyKikuchiMasterPattern, EBSDMasterPattern):
     """Lazy implementation of the :class:`EBSDMasterPattern` class.
 
     This class extends HyperSpy's LazySignal2D class for EBSD master
@@ -337,7 +242,4 @@ class LazyEBSDMasterPattern(EBSDMasterPattern, LazySignal2D):
     for attributes and methods.
     """
 
-    _lazy = True
-
-    def __init__(self, *args, **kwargs):
-        super().__init__(*args, **kwargs)
+    pass
